RF: Move TermProcessors to field.py
Tabular is the sole user of everything at this point, but there isn't
any particular reason for TermProcessors to be in tabular.py.  Group
it with its parent, StyleProcessors, instead.
